error catching in login & home
