+++ frontend/src/app/game/game-watch/game-watch.component.ts
@@ -1,7 +1,7 @@
 import { Component, ViewChild, ElementRef, AfterViewInit } from '@angular/core';
 import { Router } from '@angular/router';
 
-import { GameDisplayService } from 'src/app/services/game-data/game-display/game-display.service';
+import { GameDisplayService } from 'src/app/services/game-display/game-display.service';
 import { SocketService } from 'src/app/services/socket/socket.service';
 import { gameData } from '../game-display/GameData';
 import { User } from 'src/app/models/user';
