+++ frontend/src/app/login/login.component.css
@@ -38,3 +38,8 @@
 .pop-up-input {
 	margin-bottom: 1.5rem;
 }
+
+.red {
+	color: red;
+	margin-top: 0;
+}
